requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
backports-datetime-fromisoformat==2.0.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
distro==1.9.0
environs==9.5.0
exceptiongroup==1.3.1
fastapi==0.124.4
filelock==3.19.1
fsspec==2025.10.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
MarkupSafe==3.0.3
marshmallow==4.0.1
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
openai==2.11.0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.2
pydantic==2.12.5
pydantic_core==2.41.5
pymilvus==2.6.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.2.2
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.2
uvicorn==0.38.0
websockets==15.0.1
